Renamed some UctNode variables to be more clear. Working on the search. Specifically needs work on low memory pruning

+++ Gopher/Search.h
@@ -13,7 +13,7 @@ public:
 
 private:
 
-	int playout(Board& board, GameState& state, UctNode* const node, int depth, int color);
+	float playout(Board& board, GameState& state, UctNode* const node, int depth, int color);
 	void walkTree(Board& board, GameState& state, int& color);
 };
